requirements.txt
apache-airflow
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-snowflake==4.0.4
astronomer-providers==1.15.2
astro-run-dag
dag-factory==0.18.0
apache-airflow-providers-google==10.24.0
apache-airflow-providers-amazon
apache-airflow-providers-postgres